requirements.txt
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
PyMuPDF==1.23.8
pytesseract==0.3.10
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0